+++ crates/hdi/src/link.rs
@@ -41,12 +41,12 @@ where
 {
     fn try_into_filter(self) -> Result<LinkTypeFilter, WasmError> {
         // Collect into a 2d vector of where `LinkType`s are collected
-        // into their common `ZomeId`s.
+        // into their common `ZomeIndex`s.
         let vec = self
             .into_iter()
             .try_fold(HashMap::new(), |mut map: HashMap<_, Vec<_>>, t| {
                 let scoped = TryInto::<ScopedLinkType>::try_into(t)?;
-                map.entry(scoped.zome_id)
+                map.entry(scoped.zome_index)
                     .or_default()
                     .push(scoped.zome_type);
                 Ok(map)