+++ crates/holochain/src/conductor/api/api_cell.rs
@@ -194,7 +194,7 @@ pub trait CellConductorReadHandleT: Send + Sync {
     async fn find_cell_with_role_alongside_cell(
         &self,
         cell_id: &CellId,
-        role_id: &AppRoleId,
+        role_name: &RoleName,
     ) -> ConductorResult<Option<CellId>>;
 }
 
@@ -229,10 +229,10 @@ impl CellConductorReadHandleT for CellConductorApi {
     async fn find_cell_with_role_alongside_cell(
         &self,
         cell_id: &CellId,
-        role_id: &AppRoleId,
+        role_name: &RoleName,
     ) -> ConductorResult<Option<CellId>> {
         self.conductor_handle
-            .find_cell_with_role_alongside_cell(cell_id, role_id)
+            .find_cell_with_role_alongside_cell(cell_id, role_name)
             .await
     }
 }