+++ crates/holochain_types/src/app/app_manifest/app_manifest_validated.rs
@@ -9,7 +9,7 @@ use holochain_zome_types::DnaModifiersOpt;
 
 use super::error::{AppManifestError, AppManifestResult};
 use crate::app::app_manifest::current::{DnaLocation, DnaVersionSpec};
-use crate::prelude::AppRoleId;
+use crate::prelude::RoleName;
 use std::collections::HashMap;
 
 /// Normalized, validated representation of the App Manifest.
@@ -19,7 +19,7 @@ pub struct AppManifestValidated {
     pub(in crate::app) name: String,
 
     /// The role descriptions that make up this app.
-    pub(in crate::app) roles: HashMap<AppRoleId, AppRoleManifestValidated>,
+    pub(in crate::app) roles: HashMap<RoleName, AppRoleManifestValidated>,
 }
 
 impl AppManifestValidated {
@@ -29,13 +29,13 @@ impl AppManifestValidated {
     /// the only way to instantiate this type.
     pub(in crate::app) fn new(
         name: String,
-        roles: HashMap<AppRoleId, AppRoleManifestValidated>,
+        roles: HashMap<RoleName, AppRoleManifestValidated>,
     ) -> AppManifestResult<Self> {
-        for (role_id, role) in roles.iter() {
+        for (role_name, role) in roles.iter() {
             if let AppRoleManifestValidated::Disabled { clone_limit, .. } = role {
                 if *clone_limit == 0 {
                     return Err(AppManifestError::InvalidStrategyDisabled(
-                        role_id.to_owned(),
+                        role_name.to_owned(),
                     ));
                 }
             }