+++ crates/test_utils/wasm/wasm_workspace/countersigning/src/integrity.rs
@@ -31,12 +31,12 @@ fn validate(op: Op) -> ExternResult<ValidateCallbackResult> {
                     entry: RecordEntry::Present(entry),
                 },
         }) => {
-            match signed_action.action().entry_type().and_then(|et| match et {
-                EntryType::App(AppEntryType { id, zome_id, .. }) => Some((zome_id, id)),
+            match signed_action.action().entry_type().and_then(|entry_type| match entry_type {
+                EntryType::App(AppEntryDef { entry_index, zome_index, .. }) => Some((zome_index, entry_index)),
                 _ => None,
             }) {
-                Some((zome_id, id)) => {
-                    match EntryTypes::deserialize_from_type(*zome_id, *id, &entry) {
+                Some((zome_index, entry_index)) => {
+                    match EntryTypes::deserialize_from_type(*zome_index, *entry_index, &entry) {
                         Ok(Some(EntryTypes::Thing(thing))) => Ok(thing.into()),
                         Ok(None) => Ok(ValidateCallbackResult::Valid),
                         Err(WasmError {