+++ crates/test_utils/wasm/wasm_workspace/whoami/src/lib.rs
@@ -72,9 +72,9 @@ fn who_are_they_local(cell_id: CellId) -> ExternResult<AgentInfo> {
 }
 
 #[hdk_extern]
-fn who_are_they_role(role_id: AppRoleId) -> ExternResult<AgentInfo> {
+fn who_are_they_role(role_name: RoleName) -> ExternResult<AgentInfo> {
     let zome_call_response: ZomeCallResponse = call(
-        CallTargetCell::OtherRole(role_id),
+        CallTargetCell::OtherRole(role_name),
         zome_info()?.name,
         "whoami".to_string().into(),
         None,